Don't write an offset > size.  Add some syslog output.
This is is a workaround for issue #4.  It's should both stop the
condition, and log enough info that I should be able to figure out
the root cause the next time it happens.

There is a bunch of code that runs between when we stat the file
(to get the size) and when we dump new lines written.  If a line
is written to the log in this intervening time, it would cause
this error.